Move datatype to DAF module
