+++ src/ephemerides/translate_to_parent.rs
@@ -18,8 +18,8 @@ use crate::ephemerides::EphemInterpolationSnafu;
 use crate::hifitime::Epoch;
 use crate::math::cartesian::CartesianState;
 use crate::math::Vector3;
+use crate::naif::daf::datatypes::{HermiteSetType13, LagrangeSetType9, Type2ChebyshevSet};
 use crate::naif::daf::{DAFError, DafDataType, NAIFDataSet};
-use crate::naif::spk::datatypes::{HermiteSetType13, LagrangeSetType9, Type2ChebyshevSet};
 use crate::prelude::Frame;
 
 impl Almanac {
